+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Live daily updates of `modules.json` from the MagicMirror wiki, based on principles from [MMM-Remote-Control-Repository](https://github.com/eouia/MMM-Remote-Control-Repository).  
 - Incorporated some features found in [MMM-OnScreenMenu](https://github.com/shbatm/MMM-OnScreenMenu) that were not originally in this module.
     + Monitor (Connected Screen) On-Off Status and Toggle
+    + Delayed calls ("DELAYED" Query option and `.../delay` API paths)
     + If using Electron: Open Dev Tools, Minimize, Toggle Fullscreen
 - Configuration Option to send custom shell commands to use. Currently, only custom monitor on/off/status commands are supported.
 
@@ -30,7 +31,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Add additional MMM-OnScreenMenu features:
     + Moving modules' positions
-    + Delayed notification calls
 - PM2 Log Follower / Terminal Window
 - Added Notification Echo option to config to echo all Module Notifications to the remote's DevTools console for debugging.